moved CLI code from main.py to new class KhazadCLI
Main.py can still be used to start the app.

This change is needed for packaging the application as an AppImage.
The created AppImage contains all dependencies of the application
along with a Python distribution with the appimage module which
starts the application using an entry point in src - src:start_cli()
It expects all code to be in src.
